+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## [0.0.1]
 ### Added
 - Initial project scaffold
+
+## [1.1.0]
+### Added
+- PHP 8 Functions
+- Date functions
+- Function return types
+- Debug keywords (print_r, console, debug)
+- PHPDoc keywords